Rename spanners to legacyspanners and spanners2 to spanners
