+++ src/rendering/chord.ts
@@ -4,7 +4,7 @@ import { Note, NoteRendering } from './note';
 import { NoteDurationDenominator, StemDirection } from './enums';
 import { Clef } from './clef';
 import { KeySignature } from './keysignature';
-import { Spanners2 } from './spanners2';
+import { Spanners } from './spanners';
 
 /** The result of rendering a Chord. */
 export type ChordRendering = {
@@ -56,7 +56,7 @@ export class Chord {
   }
 
   /** Renders the Chord. */
-  render(opts: { spanners: Spanners2 }): ChordRendering {
+  render(opts: { spanners: Spanners }): ChordRendering {
     const noteRenderings = Note.render({
       notes: this.getNotes(),
       spanners: opts.spanners,
