+++ src/rendering/measure.ts
@@ -7,7 +7,7 @@ import { Config } from './config';
 import { MeasureFragment, MeasureFragmentRendering } from './measurefragment';
 import { MeasureEntry, StaveSignature } from './stavesignature';
 import { Address } from './address';
-import { Spanners2 } from './spanners2';
+import { Spanners } from './spanners';
 
 const MEASURE_LABEL_OFFSET_X = 0;
 const MEASURE_LABEL_OFFSET_Y = 24;
@@ -99,7 +99,7 @@ export class Measure {
     x: number;
     y: number;
     address: Address<'measure'>;
-    spanners: Spanners2;
+    spanners: Spanners;
     isLastSystem: boolean;
     targetSystemWidth: number;
     minRequiredSystemWidth: number;
