+++ src/rendering/voice.ts
@@ -7,7 +7,7 @@ import { GhostNote, GhostNoteRendering } from './ghostnote';
 import { Clef } from './clef';
 import { TimeSignature } from './timesignature';
 import { Address } from './address';
-import { Spanners2 } from './spanners2';
+import { Spanners } from './spanners';
 
 /** A component of a Voice. */
 export type VoiceEntry = Note | Chord | Rest | GhostNote;
@@ -58,7 +58,7 @@ export class Voice {
   }
 
   /** Renders the Voice. */
-  render(opts: { address: Address<'voice'>; spanners: Spanners2 }): VoiceRendering {
+  render(opts: { address: Address<'voice'>; spanners: Spanners }): VoiceRendering {
     const voiceEntryRenderings = this.entries.map<VoiceEntryRendering>((entry) => {
       if (entry instanceof Note) {
         return entry.render({ spanners: opts.spanners });
